api/requirements.txt
fastapi
uvicorn
mysql-connector-python
redis
pydantic
python-dotenv